fix: remove the old secret eidtor code
as we've now moved completely over to secretfacade
