--- pkg/extsecrets/editor/gsm/test_data/write_multiline/expected.json
@@ -1,1 +1,0 @@
-{"foo":"-----BEGIN PUBLIC KEY-----\nabc123\n123abc==\n-----END PUBLIC KEY-----"}
