--- pkg/extsecrets/editor/gsm/test_data/write_secrets/expected.json
@@ -1,1 +1,0 @@
-{"fish":"chips","foo":"bar","wine":"cheese"}
