--- pkg/extsecrets/editor/gsm/test_data/write_single/expected.json
@@ -1,1 +1,0 @@
-bar
